DBZ-220 Null Heartbeat object. Messages contains key
