+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/BinlogReader.java
@@ -46,7 +46,7 @@ import io.debezium.connector.mysql.MySqlConnectorConfig.EventProcessingFailureHa
 import io.debezium.connector.mysql.MySqlConnectorConfig.SecureConnectionMode;
 import io.debezium.connector.mysql.RecordMakers.RecordsForTable;
 import io.debezium.function.BlockingConsumer;
-import io.debezium.heartbeat.HeartbeatController;
+import io.debezium.heartbeat.Heartbeat;
 import io.debezium.heartbeat.OffsetPosition;
 import io.debezium.relational.TableId;
 import io.debezium.util.Clock;
@@ -86,7 +86,7 @@ public class BinlogReader extends AbstractReader {
     private final AtomicLong totalRecordCounter = new AtomicLong();
     private volatile Map<String, ?> lastOffset = null;
     private com.github.shyiko.mysql.binlog.GtidSet gtidSet;
-    private HeartbeatController heartbeat;
+    private Heartbeat heartbeat;
 
     public static class BinlogPosition {
         final String filename;
@@ -229,7 +229,8 @@ public class BinlogReader extends AbstractReader {
 
         // Set up for JMX ...
         metrics = new BinlogReaderMetrics(client);
-        heartbeat = new HeartbeatController(context.config(), context.topicSelector().getHeartbeatTopic(), () -> OffsetPosition.build(source.partition(), source.offset()));
+        heartbeat = Heartbeat.create(context.config(), context.topicSelector().getHeartbeatTopic(),
+                context.serverName(), () -> OffsetPosition.build(source.partition(), source.offset()));
     }
 
     @Override
